Rename event package to events, preventing name clashes.
