+++ events/input_file_available.go
@@ -1,4 +1,4 @@
-package event
+package events
 
 import "github.com/ONSdigital/go-ns/avro"
